Support distinct build directory from working directory
Allow to run b2 command from a configurable location, a working
directory, and pass build directory location with --build-dir option.

Add working and build directory chooser to build settings wizard.
Store working directory in BuildInfo and BuildConfiguration.
Distinguish two types of BuildStep arguments:
* additional arguments optionally specified in the BuildStepConfigWidget
* implicit arguments like the build directory selected in different
place of the wizard as it does not make sense to input in in the
"Arguments:" field. Those arguments are usually set based on
BuildConfiguration settings.